+++ Mist/Commands/Download/DownloadFirmwareOptions.swift
@@ -1,0 +1,82 @@
+//
+//  DownloadFirmwareOptions.swift
+//  Mist
+//
+//  Created by Nindi Gill on 30/5/2022.
+//
+
+import ArgumentParser
+import Foundation
+
+struct DownloadFirmwareOptions: ParsableArguments {
+
+    @Argument(help: """
+    Specify a macOS name, version or build to download:
+    * macOS Monterey
+    * macOS Big Sur
+    * macOS Catalina
+    * macOS Mojave
+    * macOS High Sierra
+    * 12.x (macOS Monterey)
+    * 11.x (macOS Big Sur)
+    * 10.15.x (macOS Catalina)
+    * 10.14.x (macOS Mojave)
+    * 10.13.x (macOS High Sierra)
+    * 21F (macOS Monterey 12.4.x)
+    * 20G (macOS Big Sur 11.6.x)
+    * 19H (macOS Catalina 10.15.7)
+    * 18G (macOS Mojave 10.14.6)
+    * 17G (macOS High Sierra 10.13.6)
+    Note: Specifying a macOS name will assume the latest version and build of that particular macOS.
+    Note: Specifying a macOS version will assume the latest build of that particular macOS.
+    """)
+    var searchString: String
+
+    @Flag(name: [.customShort("b"), .long], help: """
+    Include beta macOS Firmwares / Installers in search results.
+    """)
+    var includeBetas: Bool = false
+
+    @Flag(name: .shortAndLong, help: """
+    Force overwriting existing macOS Downloads matching the provided filename(s).
+    Note: Downloads will fail if an existing file is found and this flag is not provided.
+    """)
+    var force: Bool = false
+
+    @Option(name: .long, help: """
+    Specify the macOS Firmware output filename. The following variables will be dynamically substituted:
+    * %NAME% will be replaced with 'macOS Monterey'
+    * %VERSION% will be replaced with '12.0'
+    * %BUILD% will be replaced with '21A5304g'\n
+    """)
+    var firmwareName: String = .filenameTemplate + ".ipsw"
+
+    @Option(name: .shortAndLong, help: """
+    Specify the output directory. The following variables will be dynamically substituted:
+    * %NAME% will be replaced with 'macOS Monterey'
+    * %VERSION% will be replaced with '12.0'
+    * %BUILD% will be replaced with '21A5304g'
+    Note: Parent directories will be created automatically.\n
+    """)
+    var outputDirectory: String = .outputDirectory
+
+    @Option(name: .shortAndLong, help: """
+    Specify the temporary downloads directory.
+    Note: Parent directories will be created automatically.\n
+    """)
+    var temporaryDirectory: String = .temporaryDirectory
+
+    @Option(name: [.customShort("e"), .customLong("export")], help: """
+    Specify the path to export the download results to one of the following formats:
+    * /path/to/export.json (JSON file)
+    * /path/to/export.plist (Property List file)
+    * /path/to/export.yaml (YAML file)
+    Note: The file extension will determine the output file format.
+    """)
+    var exportPath: String?
+
+    @Flag(name: .shortAndLong, help: """
+    Suppress verbose output.
+    """)
+    var quiet: Bool = false
+}
