+++ Mist/Commands/Download/DownloadInstallerOptions.swift
@@ -1,0 +1,153 @@
+//
+//  DownloadInstallerOptions.swift
+//  Mist
+//
+//  Created by Nindi Gill on 26/8/21.
+//
+
+import ArgumentParser
+import Foundation
+
+struct DownloadInstallerOptions: ParsableArguments {
+
+    @Argument(help: """
+    Specify a macOS name, version or build to download:
+    * macOS Monterey
+    * macOS Big Sur
+    * macOS Catalina
+    * macOS Mojave
+    * macOS High Sierra
+    * 12.x (macOS Monterey)
+    * 11.x (macOS Big Sur)
+    * 10.15.x (macOS Catalina)
+    * 10.14.x (macOS Mojave)
+    * 10.13.x (macOS High Sierra)
+    * 21F (macOS Monterey 12.4.x)
+    * 20G (macOS Big Sur 11.6.x)
+    * 19H (macOS Catalina 10.15.7)
+    * 18G (macOS Mojave 10.14.6)
+    * 17G (macOS High Sierra 10.13.6)
+    Note: Specifying a macOS name will assume the latest version and build of that particular macOS.
+    Note: Specifying a macOS version will assume the latest build of that particular macOS.
+    """)
+    var searchString: String
+
+    @Argument(help: """
+    Specify the requested output type(s):
+    * application to generate a macOS Installer Application Bundle (.app).
+    * image to generate a macOS Disk Image (.dmg).
+    * iso to generate a Bootable macOS Disk Image (.iso), for use with virtualization software (ie. Parallels Desktop, VMware Fusion, VirtualBox).
+    Note: This option will fail when targeting macOS Catalina 10.15 and older on Apple Silicon (M1) Macs.
+    * package to generate a macOS Installer Package (.pkg).
+    """)
+    var outputType: [DownloadOutputType]
+
+    @Flag(name: [.customShort("b"), .long], help: """
+    Include beta macOS Firmwares / Installers in search results.
+    """)
+    var includeBetas: Bool = false
+
+    @Option(name: .shortAndLong, help: """
+    Override the default Software Update Catalog URLs.
+    """)
+    var catalogURL: String?
+
+    @Flag(name: .long, help: """
+    Cache downloaded files in the temporary downloads directory.
+    """)
+    var cacheDownloads: Bool = false
+
+    @Flag(name: .shortAndLong, help: """
+    Force overwriting existing macOS Downloads matching the provided filename(s).
+    Note: Downloads will fail if an existing file is found and this flag is not provided.
+    """)
+    var force: Bool = false
+
+    @Option(name: .long, help: """
+    Specify the macOS Installer output filename. The following variables will be dynamically substituted:
+    * %NAME% will be replaced with 'macOS Monterey'
+    * %VERSION% will be replaced with '12.0'
+    * %BUILD% will be replaced with '21A5304g'\n
+    """)
+    var applicationName: String = .filenameTemplate + ".app"
+
+    @Option(name: .long, help: """
+    Specify the macOS Disk Image output filename. The following variables will be dynamically substituted:
+    * %NAME% will be replaced with 'macOS Monterey'
+    * %VERSION% will be replaced with '12.0'
+    * %BUILD% will be replaced with '21A5304g'\n
+    """)
+    var imageName: String = .filenameTemplate + ".dmg"
+
+    @Option(name: .long, help: """
+    Codesign the exported macOS Disk Image (.dmg).
+    Specify a signing identity name, eg. "Developer ID Application: Name (Team ID)".
+    """)
+    var imageSigningIdentity: String?
+
+    @Option(name: .long, help: """
+    Specify the Bootable macOS Disk Image output filename. The following variables will be dynamically substituted:
+    * %NAME% will be replaced with 'macOS Monterey'
+    * %VERSION% will be replaced with '12.0'
+    * %BUILD% will be replaced with '21A5304g'\n
+    """)
+    var isoName: String = .filenameTemplate + ".iso"
+
+    @Option(name: .long, help: """
+    Specify the macOS Installer Package output filename. The following variables will be dynamically substituted:
+    * %NAME% will be replaced with 'macOS Monterey'
+    * %VERSION% will be replaced with '12.0'
+    * %BUILD% will be replaced with '21A5304g'\n
+    """)
+    var packageName: String = .filenameTemplate + ".pkg"
+
+    @Option(name: .long, help: """
+    Specify the macOS Installer Package identifier. The following variables will be dynamically substituted:
+    * %NAME% will be replaced with 'macOS Monterey'
+    * %VERSION% will be replaced with '12.0'
+    * %BUILD% will be replaced with '21A5304g'
+    * Spaces will be replaced with hyphens -
+    """)
+    var packageIdentifier: String?
+
+    @Option(name: .long, help: """
+    Codesign the exported macOS Installer Package (.pkg).
+    Specify a signing identity name, eg. "Developer ID Installer: Name (Team ID)".
+    """)
+    var packageSigningIdentity: String?
+
+    @Option(name: .long, help: """
+    Specify a keychain path to search for signing identities.
+    Note: If no keychain is specified, the default user login keychain will be used.
+    """)
+    var keychain: String?
+
+    @Option(name: .shortAndLong, help: """
+    Specify the output directory. The following variables will be dynamically substituted:
+    * %NAME% will be replaced with 'macOS Monterey'
+    * %VERSION% will be replaced with '12.0'
+    * %BUILD% will be replaced with '21A5304g'
+    Note: Parent directories will be created automatically.\n
+    """)
+    var outputDirectory: String = .outputDirectory
+
+    @Option(name: .shortAndLong, help: """
+    Specify the temporary downloads directory.
+    Note: Parent directories will be created automatically.\n
+    """)
+    var temporaryDirectory: String = .temporaryDirectory
+
+    @Option(name: [.customShort("e"), .customLong("export")], help: """
+    Specify the path to export the download results to one of the following formats:
+    * /path/to/export.json (JSON file)
+    * /path/to/export.plist (Property List file)
+    * /path/to/export.yaml (YAML file)
+    Note: The file extension will determine the output file format.
+    """)
+    var exportPath: String?
+
+    @Flag(name: .shortAndLong, help: """
+    Suppress verbose output.
+    """)
+    var quiet: Bool = false
+}
