+++ Mist/Commands/List/ListInstallerCommand.swift
@@ -1,84 +1,69 @@
 //
-//  List.swift
+//  ListInstallerCommand.swift
 //  Mist
 //
 //  Created by Nindi Gill on 10/3/21.
 //
 
+import ArgumentParser
 import Foundation
 import Yams
 
-/// Struct used to perform **List** operations.
-struct List {
+/// Struct used to perform **List Installer** operations.
+struct ListInstallerCommand: ParsableCommand {
+
+    static var configuration: CommandConfiguration = CommandConfiguration(
+        commandName: "installer",
+        abstract: """
+        List all macOS Installers available to download.
+        * macOS Installers for macOS Catalina 10.15 and older are for Intel based Macs only.
+        * macOS Installers for macOS Big Sur 11 and newer are Universal - for both Apple Silicon and Intel based Macs.
+        """
+    )
+    @OptionGroup var options: ListInstallerOptions
 
     /// Searches and lists the macOS versions available for download, optionally exporting to a file.
     ///
     /// - Parameters:
-    ///   - options: List options determining kind (ie. **Firmware** or **Installer**) as well as export options (ie. **CSV**, **JSON**, **Property List**, **YAML**).
+    ///   - options: List options for macOS Installers.
     ///
     /// - Throws: A `MistError` if macOS versions fail to be retrieved or exported.
-    static func run(options: ListOptions) throws {
+    static func run(options: ListInstallerOptions) throws {
         try inputValidation(options)
-
         !options.quiet ? PrettyPrint.printHeader("SEARCH") : Mist.noop()
+        !options.quiet ? PrettyPrint.print("Searching for macOS Installer versions...") : Mist.noop()
+        var catalogURLs: [String] = Catalog.urls
 
-        switch options.kind {
-        case .firmware, .ipsw:
-            !options.quiet ? PrettyPrint.print("Searching for macOS Firmware versions...") : Mist.noop()
-
-            var firmwares: [Firmware] = HTTP.retrieveFirmwares(includeBetas: options.includeBetas, quiet: options.quiet)
-
-            if let searchString: String = options.searchString {
-                firmwares = HTTP.firmwares(from: firmwares, searchString: searchString)
-            }
-
-            if options.latest {
-                if let firmware: Firmware = firmwares.first {
-                    firmwares = [firmware]
-                }
-            }
-
-            try export(firmwares.map { $0.dictionary }, options: options)
-            !options.quiet ? PrettyPrint.print("Found \(firmwares.count) macOS Firmware(s) available for download\n", prefix: .ending) : Mist.noop()
-            try list(firmwares.map { $0.dictionary }, options: options)
-
-        case .app, .installer:
-            !options.quiet ? PrettyPrint.print("Searching for macOS Installer versions...") : Mist.noop()
-
-            var catalogURLs: [String] = Catalog.urls
-
-            if let catalogURL: String = options.catalogURL {
-                catalogURLs = [catalogURL]
-            }
-
-            var products: [Product] = HTTP.retrieveProducts(from: catalogURLs, includeBetas: options.includeBetas, quiet: options.quiet)
-
-            if let searchString: String = options.searchString {
-                products = HTTP.products(from: products, searchString: searchString)
-            }
-
-            if options.latest {
-                if let product: Product = products.first {
-                    products = [product]
-                }
-            }
-
-            try export(products.map { $0.dictionary }, options: options)
-            !options.quiet ? PrettyPrint.print("Found \(products.count) macOS Installer(s) available for download\n", prefix: .ending) : Mist.noop()
-            try list(products.map { $0.dictionary }, options: options)
+        if let catalogURL: String = options.catalogURL {
+            catalogURLs = [catalogURL]
         }
+
+        var products: [Product] = HTTP.retrieveProducts(from: catalogURLs, includeBetas: options.includeBetas, quiet: options.quiet)
+
+        if let searchString: String = options.searchString {
+            products = HTTP.products(from: products, searchString: searchString)
+        }
+
+        if options.latest {
+            if let product: Product = products.first {
+                products = [product]
+            }
+        }
+
+        try export(products.map { $0.dictionary }, options: options)
+        !options.quiet ? PrettyPrint.print("Found \(products.count) macOS Installer(s) available for download\n", prefix: .ending) : Mist.noop()
+        try list(products.map { $0.dictionary }, options: options)
     }
 
     /// Perform a series of validations on input data, throwing an error if the input data is invalid.
     ///
     /// - Parameters:
-    ///   - options: List options determining kind (ie. **Firmware** or **Installer**) as well as export options (ie. **CSV**, **JSON**, **Property List**, **YAML**).
+    ///   - options: List options for macOS Installers.
     ///
     /// - Throws: A `MistError` if any of the input validations fail.
-    private static func inputValidation(_ options: ListOptions) throws {
+    private static func inputValidation(_ options: ListInstallerOptions) throws {
 
         !options.quiet ? PrettyPrint.printHeader("INPUT VALIDATION") : Mist.noop()
-        !options.quiet ? PrettyPrint.print("Kind will be '\(options.kind)'...") : Mist.noop()
 
         if let string: String = options.searchString {
 
@@ -119,10 +104,10 @@ struct List {
     ///
     /// - Parameters:
     ///   - dictionaries: The array of dictionaries to be written to disk.
-    ///   - options:      List options determining kind (ie. **Firmware** or **Installer**) as well as export options (ie. **CSV**, **JSON**, **Property List**, **YAML**).
+    ///   - options:      List options for macOS Installers.
     ///
     /// - Throws: An `Error` if the dictionaries are unable to be written to disk.
-    private static func export(_ dictionaries: [[String: Any]], options: ListOptions) throws {
+    private static func export(_ dictionaries: [[String: Any]], options: ListInstallerOptions) throws {
 
         guard let path: String = options.exportPath else {
             return
@@ -138,13 +123,7 @@ struct List {
 
         switch url.pathExtension {
         case "csv":
-            switch options.kind {
-            case .firmware, .ipsw:
-                try dictionaries.firmwaresCSVString().write(toFile: path, atomically: true, encoding: .utf8)
-            case .app, .installer:
-                try dictionaries.productsCSVString().write(toFile: path, atomically: true, encoding: .utf8)
-            }
-
+            try dictionaries.productsCSVString().write(toFile: path, atomically: true, encoding: .utf8)
             !options.quiet ? PrettyPrint.print("Exported list as CSV: '\(path)'") : Mist.noop()
         case "json":
             try dictionaries.jsonString().write(toFile: path, atomically: true, encoding: .utf8)
@@ -164,32 +143,36 @@ struct List {
     ///
     /// - Parameters:
     ///   - dictionaries: The array of dictionaries to be printed to standard output.
-    ///   - options:      List options determining kind (ie. **Firmware** or **Installer**) as well as export options (ie. **CSV**, **JSON**, **Property List**, **YAML**).
+    ///   - options:      List options for macOS Installers.
     ///
     /// - Throws: A `MistError` if the list is unable to be printed to standard output.
-    private static func list(_ dictionaries: [[String: Any]], options: ListOptions) throws {
+    private static func list(_ dictionaries: [[String: Any]], options: ListInstallerOptions) throws {
 
         switch options.outputType {
         case .ascii:
-            switch options.kind {
-            case .firmware, .ipsw:
-                print(dictionaries.firmwaresASCIIString())
-            case .app, .installer:
-                print(dictionaries.productsASCIIString())
-            }
+            print(dictionaries.productsASCIIString())
         case .csv:
-            switch options.kind {
-            case .firmware, .ipsw:
-                print(dictionaries.firmwaresCSVString())
-            case .app, .installer:
-                print(dictionaries.productsCSVString())
-            }
+            print(dictionaries.productsCSVString())
         case .json:
             try print(dictionaries.jsonString())
         case .plist:
             try print(dictionaries.propertyListString())
         case .yaml:
             try print(dictionaries.yamlString())
+        }
+    }
+
+    mutating func run() throws {
+
+        do {
+            try ListInstallerCommand.run(options: options)
+        } catch {
+            guard let mistError: MistError = error as? MistError else {
+                throw error
+            }
+
+            PrettyPrint.print(mistError.description, prefix: .ending, prefixColor: .red)
+            throw mistError
         }
     }
 }
