+++ Mist/Model/ListOutputType.swift
@@ -1,5 +1,5 @@
 //
-//  OutputType.swift
+//  ListOutputType.swift
 //  Mist
 //
 //  Created by Nindi Gill on 1/9/21.
@@ -8,7 +8,7 @@
 import ArgumentParser
 import Foundation
 
-enum OutputType: String, ExpressibleByArgument {
+enum ListOutputType: String, ExpressibleByArgument {
     // swiftlint:disable redundant_string_enum_value
     case ascii = "ascii"
     case csv = "csv"
